feat: Write UserServiceImplTest with successful registration test
This commit adds the necessary dependencies for writing unit tests, including JUnit and Mockito. It creates the UserServiceImplTest class and injects the first test method, testRegister_SuccessfulRegistration, for testing the positive scenario of user registration.

+++ src/main/java/rest/TicketBooking/model/Flight.java
@@ -3,6 +3,7 @@ package rest.TicketBooking.model;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -24,7 +25,7 @@ public class Flight extends BaseEntity{
         private LocalDateTime date_and_time_of_arrival;
 
         @Column(name = "flight_price")
-        private int flight_price;
+        private BigDecimal flight_price;
 
         @ManyToMany(fetch = FetchType.EAGER)
         @JoinTable(name = "booking",
